+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 ### Main (unreleased)
-Nil
+
+- [464](https://github.com/Shopify/job-iteration/pull/464) - Add interruption adapter for [GoodJob](https://github.com/bensheldon/good_job).
 
 ## v1.5.1 (May 29,2024)
-- [483](https://github.com/Shopify/job-iteration/pull/483) Reverts [#456 Use Arel instead of String for AR Enumerator conditionals](https://github.com/Shopify/job-iteration/pull/456)
+- [483](https://github.com/Shopify/job-iteration/pull/483) - Reverts [#456 Use Arel instead of String for AR Enumerator conditionals](https://github.com/Shopify/job-iteration/pull/456)
 
 ## v1.5.0 (May 29, 2024)
 ### Changes